+++ catalogue/retrywrappers/PhysicalLibraryCatalogueRetryWrapper.hpp
@@ -52,26 +52,8 @@ public:
 
   std::list<common::dataStructures::PhysicalLibrary> getPhysicalLibraries() const override;
 
-  void modifyPhysicalLibraryGuiUrl   (const common::dataStructures::SecurityIdentity &admin,
-   const std::string &name, const std::string &guiUrl) override;
-
-  void modifyPhysicalLibraryWebcamUrl(const common::dataStructures::SecurityIdentity &admin,
-   const std::string &name, const std::string &webcamUrl) override;
-
-  void modifyPhysicalLibraryLocation (const common::dataStructures::SecurityIdentity &admin,
-   const std::string &name, const std::string &location) override;
-
-  void modifyPhysicalLibraryNbPhysicalCartridgeSlots  (const common::dataStructures::SecurityIdentity &admin,
-   const std::string &name, const uint64_t &nbPhysicalCartridgeSlots) override;
-
-  void modifyPhysicalLibraryNbAvailableCartridgeSlots (const common::dataStructures::SecurityIdentity &admin,
-   const std::string &name, const uint64_t &nbAvailableCartridgeSlots) override;
-
-  void modifyPhysicalLibraryNbPhysicalDriveSlots (const common::dataStructures::SecurityIdentity &admin,
-   const std::string &name, const uint64_t &nbPhysicalDriveSlots) override;
-
-  void modifyPhysicalLibraryComment (const common::dataStructures::SecurityIdentity &admin,
-   const std::string &name, const std::string &comment) override;
+  void modifyPhysicalLibrary(const common::dataStructures::SecurityIdentity& admin,
+    const common::dataStructures::UpdatePhysicalLibrary& pl) override;
 
 private:
   const std::unique_ptr<Catalogue>& m_catalogue;